refactored README.md and built out the query APIs for Cassandra
