+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "test": "jest",
     "build": "webpack",
-    "start": "node ./server/index.js",
-    "seed": "node ./server/db/seed.js"
+    "start": "nodemon ./server/index.js",
+    "seed": "node ./server/db/seeding.js"
   },
   "repository": {
     "type": "git",
@@ -25,6 +25,7 @@
     "@testing-library/react": "^9.4.0",
     "axios": "^0.19.2",
     "body-parser": "^1.19.0",
+    "cassandra-driver": "^4.4.0",
     "cors": "^2.8.5",
     "csv-writer": "^1.6.0",
     "express": "^4.17.1",
