+++ server/db/index.js
@@ -1,17 +1,9 @@
-const mongoose = require('mongoose');
-// const mongoUrl = 'mongodb://database/gallery';
+const cassandra = require('cassandra-driver');
 
-// mongoose.connect(mongoUrl, { server: { reconnectTries: Number.MAX_VALUE } })
-//   .then(() => {
-//     console.log('Successfully connected to MongoDb');
-//   })
-//   .catch((error) => console.log(error));
-
-
-const GallerySchema = mongoose.Schema({
-  listing_id: { type: Number, required: true, unique: true },
-  listing_title: { type: String },
-  listing_images: [{ id: Number, url: String, caption: String }],
+const client = new cassandra.Client({
+  contactPoints: ['127.0.0.1'],
+  localDataCenter: 'datacenter1',
+  keyspace: 'sdc_airbusybeavers'
 });
 
-module.exports = mongoose.model('Gallery', GallerySchema);
+module.exports = client;
