+++ server/index-mongo.js
@@ -16,14 +16,13 @@ app.use(express.static(path.join(__dirname, "../client/dist")));
 
 app.get('/gallery/:id', controllers.gallery.getOne);
 
-app.get('/galleries', controllers.gallery.getAll);
+app.get('/gallery', controllers.gallery.getAll);
 
 app.post('/gallery', controllers.gallery.postOne);
 
-app.put('/gallery/:id', controllers.gallery.updateOne);
-
 app.delete('/gallery/:id', controllers.gallery.deleteOne);
 
-app.delete('/gallery/', controllers.gallery.deleteAll);
+app.put('/gallery/:id', controllers.gallery.updateOne);
+
 
 app.listen(port, () => console.log(`Server listening on port ${port}!`));
