+++ server/models/index.js
@@ -1,80 +1,78 @@
-const mongoose = require('mongoose');
-const mongoUrl = 'mongodb://localhost/gallery';
-const GalleryModel = require('../db');
-
-mongoose.connect(mongoUrl, { server: { reconnectTries: Number.MAX_VALUE } });
+const client = require('../db');
 
 module.exports = {
   gallery: {
-    insertAll: (data, cb) => {
-      GalleryModel.insertMany(data)
-        .then((docs) => {
-          cb(null, docs);
-        })
-        .catch((err) => {
-          cb(err);
-        });
-    },
 
     getOne: (id, cb) => {
-      GalleryModel.find({ listing_id: Number(id) })
-        .then((results) => {
-          cb(null, results);
+      var getOneListing = 'SELECT * FROM nosqldata WHERE listing_id = ?';
+
+      client.execute(getOneListing, [id], { prepare: true })
+        .then((result) => {
+          cb(null, result.rows);
         })
-        .catch((err) => {
-          cb(err);
-        });
+        .catch((error) => {
+          cb(error);
+        })
     },
 
     getAll: (cb) => {
-      GalleryModel.find({})
-        .then((results) => {
-          cb(null, results);
+      var getAllListings = 'SELECT * FROM nosqldata';
+
+      client.execute(getAllListings)
+        .then((result) => {
+          cb(null, result.rows);
         })
-        .catch((err) => {
-          cb(err);
-        });
+        .catch((error) => {
+          cb(error);
+        })
     },
 
     postOne: (content, cb) => {
-      var NewGallery = new GalleryModel(content);
-      NewGallery.save()
-        .then((results) => {
-          cb(null, results);
+      var insertListing = 'INSERT INTO nosqldata(listing_id, listing_title, id, url, caption) VALUES (?, ?, ?, ?, ?)';
+
+      client.execute(insertListing, [content.listing_id, content.listing_title, content.id, content.url, content.caption], { prepare: true })
+        .then(() => {
+          cb(null, 'success');
         })
-        .catch((err) => {
-          cb(err);
-        });
+        .catch((error) => {
+          cb(error);
+        })
     },
 
     deleteOne: (id, cb) => {
-      GalleryModel.deleteOne({listing_id: Number(id)})
-        .then((results) => {
-          cb(null, results);
+      var deleteOne = 'DELETE FROM nosqldata WHERE listing_id = ?';
+
+      client.execute(deleteOne, [id], { prepare: true })
+        .then(() => {
+          cb(null, 'delete success');
         })
-        .catch((err) => {
-          cb(err);
+        .catch((error) => {
+          cb(error);
         })
     },
 
     updateOne: (id, content, cb) => {
-      GalleryModel.updateOne({listing_id: Number(id)}, content)
-        .then((results) => {
-          cb(null, results);
-        })
-        .catch((err) => {
-          cb(err);
-        })
+      var updateOne = 'UPDATE nosqldata SET listing_title = ?, url = ?, caption = ? WHERE listing_id = ? and id = ?';
+
+      client.execute(updateOne, [content.listing_title, content.url, content.caption, id, content.id], { prepare: true })
+      .then(() => {
+        cb(null, 'update success');
+      })
+      .catch((error) => {
+        cb(error);
+      })
     },
 
     deleteAll: (cb) => {
-      GalleryModel.deleteMany({})
-        .then((results) => {
-          cb(null, results);
+      var deleteAll = 'TRUNCATE nosqldata';
+
+      client.execute(deleteAll)
+        .then(() => {
+          cb(null, 'delete all success');
         })
         .catch((err) => {
           cb(err);
-        });
-    },
-  },
+        })
+    }
+  }
 };
